refactor(agent): modularize mirror lifecycle
