+++ config/architecture-budgets.json
@@ -4,7 +4,6 @@
     "apps/desktop/src/main/main.ts": 1022,
     "apps/desktop/src/renderer/main.tsx": 1714,
     "apps/portal/src/main.tsx": 2251,
-    "crates/connect-agent/src/mirrors.rs": 1594,
     "services/server/src/legacy-baseline.ts": 1114
   }
 }
